Use declarative approach in all modules
All modules are now using a more-declarative approach in which all the
information to create the whole section (eg: ingredients list, method
steps etc) is being returned as an object by the module, and is then
being rendered in recipe.js, with the parent element being passed in.

There is one exception; metadata.js, which still renders more than once
within the module. This is due to some complexity with sibling elements
and will be addressed in a future PR.